Refactor CLI as a submodule with cli.py broken into smaller modules for each subcommand.
